+++ Runtime/Scripts/IPathfinding.cs
@@ -24,29 +24,6 @@ namespace Kokowolo.Pathfinding
         /*██████████████████████████████████████████████████████████*/
         #region Functions
 
-        // NOTE: this may change based on pathfinding configuration, i.e. a node's unit using its GetNeighbors() vs what's below
-        public List<INode> GetNeighborsFromNode(INode iNode)
-        {
-            return iNode.Node.GetNeighbors();
-        }
-
-        public bool IsValidMoveBetweenNodes(INode start, INode end)
-        {
-            return start.Node.HasNeighbor(end);
-        }
-
-        // NOTE: this should get the distance/approximate between these two nodes
-        public int GetHeuristicCostBetweenNodes(INode start, INode end);
-
-        // NOTE: this should get the actual distance between these two nodes
-        public int GetMoveCostBetweenNodes(INode start, INode end);
-
-        // NOTE: this method is a hacky way of checking if the current path is too long; i.e. path.Distance > maxDistance;
-        public bool IsPathOutsideMovementRange(INodePath path)
-        {
-            return false;
-        }
-
         // TODO: add burst or jobs to package
         // internal void OnSearchComplete(PathfindingNodePath path);