+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.20] - 2026-01-25
+### Added
+* `INodeComposition` to have an interface for handling a `Node` object
+* `IPathfinder` splitting up functionality of `IPathfinding` to a manager interface and an instance interface
+### Changed
+* AStarPathfinding to use `INode` without a `INode.Node` member
+* `INode` to have all properties that `Node` had and `Node` to serve as an object containing `INode` properties
+* all classes starting with `INode`[...] to now lead with `Node`[...]
+### Removed
+* all confusing neighbor functions within `Node`
+
 ## [0.0.19] - 2026-01-07
 ### Changed
 * all of `AStarPathfinding` to use `INode` instead of `Node`